First tests for classes and added compatibility with master branch
