+++ src/gamecontroller/gamecontrollerset.h
@@ -20,8 +20,12 @@
 
 #include "setjoystick.h"
 
-#include <QObject>
+#include <SDL2/SDL_gamecontroller.h>
 
+#include <QObject>
+#include <QHash>
+
+class QXmlStreamReader;
 class InputDevice;
 
 
@@ -34,6 +38,11 @@ public:
 
     virtual void refreshAxes();
 
+    virtual void readConfig(QXmlStreamReader *xml); // GameControllerSetXml
+    virtual void readJoystickConfig(QXmlStreamReader *xml,
+                            QHash<int, SDL_GameControllerButton> &buttons,
+                            QHash<int, SDL_GameControllerAxis> &axes,
+                            QList<SDL_GameControllerButtonBind> &hatButtons); // GameControllerSetXml class
 
 protected:
     void populateSticksDPad();
@@ -42,8 +51,9 @@ public slots:
     virtual void reset();
 
 private:
+    void getElemFromXml(QString elemName, QXmlStreamReader *xml); // GameControllerSetXml class
+    void readConfDpad(QXmlStreamReader *xml, QList<SDL_GameControllerButtonBind> &hatButtons, bool vdpadExists, bool dpadExists); // GameControllerSetXml class
     void resetSticks();
-
 };
 
 #endif // GAMECONTROLLERSET_H
