+++ src/gamecontroller/xml/gamecontrollerxml.cpp
@@ -1,7 +1,6 @@
 #include "gamecontrollerxml.h"
 
 #include "gamecontroller/gamecontroller.h"
-#include "gamecontrollersetxml.h"
 
 #include "globalvariables.h"
 #include "common.h"
@@ -60,8 +59,7 @@ void GameControllerXml::readJoystickConfig(QXmlStreamReader *xml)
                         if ((index >= 0) && (index < m_gameController->getJoystick_sets().size()))
                         {
                             GameControllerSet *currentSet = qobject_cast<GameControllerSet*>(m_gameController->getJoystick_sets().value(index)); // static_cast
-                            GameControllerSetXml *gameContrSet = new GameControllerSetXml(currentSet);
-                            gameContrSet->readJoystickConfig(xml, buttons, axes, hatButtons);
+                            currentSet->readJoystickConfig(xml, buttons, axes, hatButtons);
                         }
                     }
                     else
@@ -307,8 +305,7 @@ void GameControllerXml::readConfig(QXmlStreamReader *xml)
 
                         if ((index >= 0) && (index < m_gameController->getJoystick_sets().size()))
                         {
-                            SetJoystickXml *setJoystickXml = new SetJoystickXml(m_gameController->getJoystick_sets().value(index));
-                            setJoystickXml->readConfig(xml);
+                            m_gameController->getJoystick_sets().value(index)->readConfig(xml);
                         }
                     }
                     else
@@ -434,8 +431,7 @@ void GameControllerXml::writeConfig(QXmlStreamWriter *xml)
 
     for (int i = 0; i < m_gameController->getJoystick_sets().size(); i++)
     {
-        SetJoystickXml *setJoystickXml = new SetJoystickXml(m_gameController->getJoystick_sets().value(i));
-        setJoystickXml->writeConfig(xml);
+        m_gameController->getJoystick_sets().value(i)->writeConfig(xml);
     }
 
     xml->writeEndElement();
