+++ src/inputdevice.cpp
@@ -29,6 +29,8 @@
 
 #include <typeinfo>
 
+#include <QXmlStreamReader>
+#include <QXmlStreamWriter>
 #include <QDebug>
