+++ src/inputdevice.h
@@ -28,6 +28,9 @@
 #include <QRegExp>
 
 class AntiMicroSettings;
+class SetJoystick;
+class QXmlStreamReader;
+class QXmlStreamWriter;
 
 
 class InputDevice : public QObject
