+++ src/joyaxis.cpp
@@ -27,6 +27,8 @@
 
 #include <cmath>
 
+#include <QXmlStreamReader>
+#include <QXmlStreamWriter>
 #include <QtGlobal>
 #include <QDebug>
 
@@ -679,10 +681,11 @@ void JoyAxis::removeControlStick(bool performRelease)
 
     if (m_stick != nullptr)
     {
-        if (performRelease)
+        if (performRelease) {
             m_stick->releaseButtonEvents();
+            m_stick = nullptr;
+        }
 
-        m_stick = nullptr;
         emit propertyUpdated();
     }
 }
